(RK-357) Restrict `gettext` and `fast_gettext` versions
This commit restricts the allowed versions of the `gettext` and
`fast_gettext` gems to preserve compatability with older versions of
Ruby. This became necessary because the `gettext-setup` gem loosened its
requirements to allow newer versions of those deps, as part of a
modernization effort. We can remove these pins once we only need to
support Ruby 2.5+.