[#1106, #1109] Save credentials from CLI flags
This commit adds a step to the `deploy environment` command invocation
that takes the credential files as passed via the new flags, reads their
contents, and saves it to the settings hash for the deploy action. This
hash is passed through to the deployment object, which is responsible
for performing the bulk of the action logic.